fixed typos

+++ scripts/checkoutEnterprise.fish
@@ -3,7 +3,7 @@ eval $SCRIPTSDIR/checkoutArangoDB.fish
 and cd $INNERWORKDIR/ArangoDB
 and if test ! -d enterprise
   git clone ssh://git@github.com/arangodb/enterprise
-else test ! -d enterprise/.git
+else if test ! -d enterprise/.git
   rm -rf enterprise
   git clone ssh://git@github.com/arangodb/enterprise
 end
